requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
scikit-learn
folium
haversine
Pillow
jinja2